Fixes #16752 - Extract template code from Host::Managed
Host::Managed contains a lot of code related with determining the
template that a host should have. Other Host objects such as
Host::Discovered might need that logic, without becoming a
Host::Managed.

This commit extracts that into Hostext::OperatingSystem so it can
be used on plugins.
